+++ app/src/main/java/org/simple/clinic/medicalhistory/MedicalHistoryQuestion.kt
@@ -22,4 +22,5 @@ sealed class MedicalHistoryQuestion(
   )
 
   object IsOnDiabetesTreatment : MedicalHistoryQuestion(R.string.medicalhistory_question_is_on_diabetes_treatment)
+  data object IsSmoking : MedicalHistoryQuestion(R.string.medicalhistory_question_is_current_smoker)
 }