+++ app/src/main/java/org/simple/clinic/summary/medicalhistory/MedicalHistorySummaryModel.kt
@@ -13,12 +13,14 @@ import java.util.UUID
 data class MedicalHistorySummaryModel(
     val patientUuid: UUID,
     val medicalHistory: MedicalHistory? = null,
-    val currentFacility: Facility? = null
+    val currentFacility: Facility? = null,
+    val showIsSmokingQuestion: Boolean,
 ) : Parcelable, PatientSummaryChildModel {
 
   companion object {
-    fun create(patientUuid: UUID): MedicalHistorySummaryModel = MedicalHistorySummaryModel(
-        patientUuid = patientUuid
+    fun create(patientUuid: UUID, showIsSmokingQuestion: Boolean): MedicalHistorySummaryModel = MedicalHistorySummaryModel(
+        patientUuid = patientUuid,
+        showIsSmokingQuestion = showIsSmokingQuestion
     )
   }