+++ app/src/test/java/org/simple/clinic/medicalhistory/newentry/NewMedicalHistoryScreenLogicTest.kt
@@ -1,5 +1,11 @@
 package org.simple.clinic.medicalhistory.newentry
 
+import io.reactivex.Completable
+import io.reactivex.rxkotlin.ofType
+import io.reactivex.subjects.PublishSubject
+import org.junit.After
+import org.junit.Rule
+import org.junit.Test
 import org.mockito.kotlin.any
 import org.mockito.kotlin.eq
 import org.mockito.kotlin.inOrder
@@ -7,13 +13,6 @@ import org.mockito.kotlin.mock
 import org.mockito.kotlin.times
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
-import dagger.Lazy
-import io.reactivex.Completable
-import io.reactivex.rxkotlin.ofType
-import io.reactivex.subjects.PublishSubject
-import org.junit.After
-import org.junit.Rule
-import org.junit.Test
 import org.simple.clinic.appconfig.Country
 import org.simple.clinic.medicalhistory.Answer.No
 import org.simple.clinic.medicalhistory.Answer.Unanswered
@@ -25,6 +24,7 @@ import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HasHadAKidneyDise
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.HasHadAStroke
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.IsOnDiabetesTreatment
 import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.IsOnHypertensionTreatment
+import org.simple.clinic.medicalhistory.MedicalHistoryQuestion.IsSmoking
 import org.simple.clinic.medicalhistory.MedicalHistoryRepository
 import org.simple.clinic.medicalhistory.OngoingMedicalHistoryEntry
 import org.simple.clinic.patient.Gender
@@ -121,6 +121,7 @@ class NewMedicalHistoryScreenLogicTest {
     uiEvents.onNext(NewMedicalHistoryAnswerToggled(HasHadAKidneyDisease, Yes))
     uiEvents.onNext(NewMedicalHistoryAnswerToggled(DiagnosedWithDiabetes, Yes))
     uiEvents.onNext(NewMedicalHistoryAnswerToggled(IsOnDiabetesTreatment, Yes))
+    uiEvents.onNext(NewMedicalHistoryAnswerToggled(IsSmoking, Yes))
     uiEvents.onNext(SaveMedicalHistoryClicked())
 
     // then
@@ -145,7 +146,8 @@ class NewMedicalHistoryScreenLogicTest {
               hasHadKidneyDisease = Yes,
               diagnosedWithHypertension = No,
               hasDiabetes = Yes,
-              isOnDiabetesTreatment = Yes
+              isOnDiabetesTreatment = Yes,
+              isSmoking = Yes,
           )
       )
       verify(uiActions).openPatientSummaryScreen(savedPatient.patientUuid)
@@ -209,7 +211,9 @@ class NewMedicalHistoryScreenLogicTest {
               hasHadKidneyDisease = Unanswered,
               diagnosedWithHypertension = Yes,
               hasDiabetes = Unanswered,
-              isOnHypertensionTreatment = Yes))
+              isOnHypertensionTreatment = Yes,
+              isSmoking = Unanswered,
+          ))
       verify(uiActions).openPatientSummaryScreen(savedPatient.patientUuid)
     }
   }
@@ -248,6 +252,7 @@ class NewMedicalHistoryScreenLogicTest {
     uiEvents.onNext(NewMedicalHistoryAnswerToggled(HasHadAStroke, No))
     uiEvents.onNext(NewMedicalHistoryAnswerToggled(HasHadAKidneyDisease, Yes))
     uiEvents.onNext(NewMedicalHistoryAnswerToggled(DiagnosedWithDiabetes, Yes))
+    uiEvents.onNext(NewMedicalHistoryAnswerToggled(IsSmoking, Yes))
 
     // Updated answers
     uiEvents.onNext(NewMedicalHistoryAnswerToggled(DiagnosedWithHypertension, Yes))
@@ -256,6 +261,7 @@ class NewMedicalHistoryScreenLogicTest {
     uiEvents.onNext(NewMedicalHistoryAnswerToggled(HasHadAKidneyDisease, No))
     uiEvents.onNext(NewMedicalHistoryAnswerToggled(DiagnosedWithDiabetes, No))
     uiEvents.onNext(NewMedicalHistoryAnswerToggled(IsOnHypertensionTreatment(Country.INDIA), Yes))
+    uiEvents.onNext(NewMedicalHistoryAnswerToggled(IsSmoking, No))
 
     uiEvents.onNext(SaveMedicalHistoryClicked())
 
@@ -281,7 +287,8 @@ class NewMedicalHistoryScreenLogicTest {
               hasHadKidneyDisease = No,
               diagnosedWithHypertension = Yes,
               hasDiabetes = No,
-              isOnHypertensionTreatment = Yes
+              isOnHypertensionTreatment = Yes,
+              isSmoking = No,
           )
       )
       verify(uiActions).openPatientSummaryScreen(savedPatient.patientUuid)
@@ -302,8 +309,8 @@ class NewMedicalHistoryScreenLogicTest {
         patientRepository = patientRepository,
         medicalHistoryRepository = medicalHistoryRepository,
         dataSync = mock(),
-        currentUser = Lazy { user },
-        currentFacility = Lazy { facility },
+        currentUser = { user },
+        currentFacility = { facility },
         uuidGenerator = uuidGenerator,
         dateOfBirthFormatter = dateOfBirthFormatter,
         viewEffectsConsumer = viewEffectHandler::handle
@@ -313,7 +320,7 @@ class NewMedicalHistoryScreenLogicTest {
 
     testFixture = MobiusTestFixture(
         events = uiEvents.ofType(),
-        defaultModel = NewMedicalHistoryModel.default(country),
+        defaultModel = NewMedicalHistoryModel.default(country, true),
         init = NewMedicalHistoryInit(),
         update = NewMedicalHistoryUpdate(),
         effectHandler = effectHandler,